requirements.txt
requests>=2.28
APScheduler>=3.10
python-dotenv>=1.0
